Fix environment variables and integrate authentication flow
- Renamed GEMINI_API_KEY and DEEPGRAM_API_KEY to REACT_APP_ prefix in .env, src/nlp/reframe-engine.js, and src/ui/App.jsx.
- Fixed syntax error in src/ui/Login.jsx by moving import statement to top.
- Refactored src/ui/App.jsx to use react-router-dom for routing between Login, Signup, and Dashboard.
- Extracted main application logic from App.jsx to new src/ui/Dashboard.jsx component.
- Added Logout functionality to Dashboard.
- Enabled JWT token interceptor in src/services/api.js.
- Updated src/nlp/cd-detector.js to use manual dot product for cosine similarity, avoiding deprecated tf.metrics.cosineDistance.
- Removed webpack-dev-server override from package.json to fix react-scripts compatibility.
- Updated .gitignore to exclude build artifacts and logs.
