+++ .gitignore
@@ -1,2 +1,7 @@
 node_modules
 .jules_api_key
+.env
+build/
+*.log
+verification/
+__pycache__/
